+++ CHANGELOG.md
@@ -20,6 +20,13 @@
 </p>
 
 ---
+### 2025.12.0.dev6 / Development Release
+**Changes:**
+- `neuronum serve-agent` is no longer part of the SDK 
+- `neuronum stop-agent` is no longer part of the SDK 
+- `setup.sh` is now start_neuronum_server.sh
+- use `stop_neuronum_server.sh` to stop the server
+- .env credentials replace MNEMONIC value in server.config 
 
 ### 2025.12.0.dev5 / Development Release
 **Changes:**